+++ src/X.Web.MRSS/Structure/MediaCategory.cs
@@ -1,0 +1,22 @@
+using JetBrains.Annotations;
+
+namespace X.Web.MRSS.Structure;
+
+[PublicAPI]
+public record MediaCategory
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public string? Scheme { get; set; }
+    
+    /// <summary>
+    /// 
+    /// </summary>
+    public string? Label { get; set; }
+    
+    /// <summary>
+    /// e.g., "music/artist/album/song"
+    /// </summary>
+    public string? Value { get; set; }
+}
