+++ src/X.Web.MRSS/Structure/MediaRssItem.cs
@@ -79,79 +79,3 @@ public record MediaRssItem : RssItem
 
     public List<MediaCredit> Credits { get; set; } = new();
 }
-
-[PublicAPI]
-public record MediaRating
-{
-    /// <summary>
-    /// e.g., "urn:simple"
-    /// </summary>
-    public string? Scheme { get; set; }
-    
-    /// <summary>
-    /// e.g., "adult", "nonadult"
-    /// </summary>
-    public string? Value { get; set; }
-}
-
-[PublicAPI]
-public record MediaThumbnail
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public string? Url { get; set; }
-    
-    /// <summary>
-    /// 
-    /// </summary>
-    public int Width { get; set; }
-    
-    /// <summary>
-    /// 
-    /// </summary>
-    public int Height { get; set; }
-    
-    /// <summary>
-    /// Format "H:M:S.h" or "S.h"
-    /// </summary>
-    public string Time { get; set; }
-}
-
-[PublicAPI]
-public record MediaCategory
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public string? Scheme { get; set; }
-    
-    /// <summary>
-    /// 
-    /// </summary>
-    public string? Label { get; set; }
-    
-    /// <summary>
-    /// e.g., "music/artist/album/song"
-    /// </summary>
-    public string? Value { get; set; }
-}
-
-[PublicAPI]
-public record MediaCredit
-{
-    /// <summary>
-    /// e.g., "producer", "owner"
-    /// </summary>
-    public string? Role { get; set; }
-    
-    /// <summary>
-    /// e.g., "urn:ebu"
-    /// </summary>
-    public string? Scheme { get; set; }
-    
-    /// <summary>
-    /// 
-    /// </summary>
-    public string? Value { get; set; }
-}
